pyproject.toml
[tool.poetry]
name = "chatgpt-discord"
version = "0.1.0"
description = ""
authors = ["hyeonproject <axzswq@gmail.com>"]
readme = "README.md"
packages = [{include = "chatgpt_discord"}]

[tool.poetry.dependencies]
python = "^3.11"
discordwebhook = "^1.0.3"
fastapi = "^0.93.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
httpx = "^0.23.3"
pymongo = "^4.3.3"
websockets = "^10.4"
python-multipart = "^0.0.6"
discord-py = "^2.2.2"
openai = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
